% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{capso_get_theme}
\alias{capso_get_theme}
\title{Gets the capso theme for ggplot}
\usage{
capso_get_theme()
}
\value{
The theme object for the custom ggplot2 theme.
}
\description{
Gets the capso theme for ggplot
}
