% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{capso_phase_plot}
\alias{capso_phase_plot}
\title{Plots the phase plot of two time series contained in parameter data}
\usage{
capso_phase_plot(data, title = "", xlabel = "", ylabel = "")
}
\arguments{
\item{data}{A two-column matrix containing two time series.}

\item{title}{The title for the plot.}

\item{xlabel}{The label for the x axis.}

\item{ylabel}{The label for the y axis.}
}
\description{
Plots the phase plot of two time series contained in parameter data
}
