% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rcapso_adjust_parameters}
\alias{rcapso_adjust_parameters}
\title{Adjusts the parameters of the mean field model using simulation data}
\usage{
rcapso_adjust_parameters(df, transient_length = 1000)
}
\arguments{
\item{df}{A data frame containing normalize data of a CaPso simulation.}

\item{transient_length}{The length (in seasons) of the transient.}
}
\value{
A numerical vector containing the adjusted parameters of the mean
field model.
}
\description{
Adjusts the parameters of the mean field model using simulation data
}
