% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rcapso_get_theme}
\alias{rcapso_get_theme}
\title{Gets the rcapso theme for ggplot}
\usage{
rcapso_get_theme()
}
\value{
The ggplot2 theme object for the rcapso package.
}
\description{
Gets the rcapso theme for ggplot
}
