% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rcapso_mean_field}
\alias{rcapso_mean_field}
\title{Gets the mean field approximation of the CAPSO model.}
\usage{
rcapso_mean_field(
  use_reg = TRUE,
  num_iter = 100,
  psi0 = 1,
  phi0 = 0.01,
  alpha = 0.1,
  ey = 1,
  ry = 1,
  ez = 1,
  rz = 1,
  a = -1,
  b = 1,
  d = 1,
  e = 0
)
}
\arguments{
\item{use_reg}{Indicates if regression terms should be used.}

\item{num_iter}{The number of iterations to simulate.}

\item{psi0}{The initial density of the prey population.}

\item{phi0}{The initial density of the predator population.}

\item{alpha}{The intraspecific competition coefficient.}

\item{ey}{The reproductive capacity of preys.}

\item{ry}{The radius of the prey's reproduction neighborhood.}

\item{ez}{The reproductive capacity of predators.}

\item{rz}{The radius of the predator's reproduction neighborhood.}

\item{a}{The coefficient of the line that defines predator mortality.}

\item{b}{The intercept of the line that defines predator mortality.}

\item{d}{The coefficient of the line that defines prey mortality.}

\item{e}{The intercept of the line that defines prey mortality.}
}
\value{
A 2-column data frame containing the time series of preys and
predators.
}
\description{
Gets the mean field approximation of the CAPSO model.
}
