% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rcapso_mean_field_growth_rate}
\alias{rcapso_mean_field_growth_rate}
\title{Calculates the mean growth rate of a population}
\usage{
rcapso_mean_field_growth_rate(n = 100, epsilon = 1, radius = 1)
}
\arguments{
\item{n}{The length of the output vector.}

\item{epsilon}{The reproductive capacity of the population.}

\item{radius}{The reproductive capacity of the population.}
}
\value{
A numerical vector having the mean growth rate for n input values
between 0 and 1 values.
}
\description{
Calculates the mean growth rate of a population
}
