% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rcapso_plot_fourier_spectrum}
\alias{rcapso_plot_fourier_spectrum}
\title{Plots the Fourier spectrum of a time series.}
\usage{
rcapso_plot_fourier_spectrum(x, method_spec)
}
\arguments{
\item{x}{A numeric vector containing the data of the time series.}

\item{method_spec}{String specifying the method used to estimate the
spectral density.}
}
\value{
A list containing the cycles sorted in order of magnitude of the
power spikes.
}
\description{
Plots the Fourier spectrum of a time series.
}
