% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rcapso_plot_phase}
\alias{rcapso_plot_phase}
\title{Plots the phase plot of two time series contained in parameter data}
\usage{
rcapso_plot_phase(
  data,
  cols = c("Preys", "Predators"),
  title = "Phase Plot",
  normalize = TRUE,
  lat_size = 262144
)
}
\arguments{
\item{data}{A data.frame containing the time series whose phase plot should
be obtained.}

\item{cols}{The columns in the dataframe that contain the
time series to be processed.}

\item{title}{The title for the plot.}

\item{normalize}{TRUE if the data needs to be converted to a density.}

\item{lat_size}{The size of the lattice (only used if normalize is TRUE).}
}
\description{
Plots the phase plot of two time series contained in parameter data
}
