% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rcapso_plot_prey_pred_data}
\alias{rcapso_plot_prey_pred_data}
\title{Plots the time series of preys and predators.}
\usage{
rcapso_plot_prey_pred_data(data, title = "", xlabel = "", ylabel = "")
}
\arguments{
\item{data}{A two-column matrix containing the data of preys and predators.}

\item{title}{The title for the plot.}

\item{xlabel}{The label for the x axis.}

\item{ylabel}{The label for the y axis.}
}
\description{
Plots the time series of preys and predators.
}
