% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rcapso_mean_field_growth_rate}
\alias{rcapso_mean_field_growth_rate}
\title{Calculates the mean field growth rate of a population}
\usage{
rcapso_mean_field_growth_rate(pop_data, epsilon = 1, radius = 1)
}
\arguments{
\item{pop_data}{A numeric vector containing populatin density data.}

\item{epsilon}{The reproductive capacity of the population.}

\item{radius}{The reproductive capacity of the population.}
}
\value{
A numerical vector with the mean growth rate for every value in
pop_data.
}
\description{
Calculates the mean field growth rate of a population
}
